updated readme file with better instructions, fixed typescript and lint issues
