+++ @types/alltypes.d.ts
@@ -1,4 +1,13 @@
-import * as React from "react";
+declare module 'hoist-non-react-statics' {
+  import {ComponentClass} from 'react';
+
+  function hoistNonReactStatics<P>(to: ComponentClass<P>, from: ComponentClass<any>): ComponentClass<P>;
+  export = hoistNonReactStatics;
+}
+
+declare module 'rc-form';
+
+declare module 'react-addons-perf';import * as React from "react";
 
 declare module "react-css-themr" {
   type TReactCSSThemrTheme = {
@@ -32,4 +41,9 @@ declare module "react-css-themr" {
     defaultTheme?: {},
     options?: IThemrOptions
   ): <P, S>(component: (new(props?: P, context?: any) => React.Component<P, S>) | React.SFC<P>) => ThemedComponentClass<P, S>;
-}
+}declare module 'react-page-layout';
+
+declare module 'react-perf-tool';
+declare module 'react-prism';
+declare module 'react-router-dom';
+declare module 'velocity-animate';
