--- @types/rc-form.d.ts
@@ -1,4 +1,0 @@
-declare module 'rc-form';
-declare module 'react-addons-perf';
-declare module 'react-perf-tool';
-declare module 'react-page-layout';
