--- config/typescript/modules.d.ts
@@ -1,11 +1,0 @@
-declare module 'hoist-non-react-statics' {
-  import {ComponentClass} from 'react';
-
-  function hoistNonReactStatics<P>(to: ComponentClass<P>, from: ComponentClass<any>): ComponentClass<P>;
-  export = hoistNonReactStatics;
-}
-
-declare module 'velocity-animate';
-declare module 'react-router-dom';
-declare module 'react-prism';
-declare module 'react-table';
